Move text cleaning to utils
